+++ demisto_sdk/commands/common/hook_validations/docker.py
@@ -43,6 +43,10 @@ class DockerImageValidator(BaseValidator):
         self.docker_image_name, self.docker_image_tag = self.parse_docker_image(self.yml_docker_image)
         self.is_latest_tag = True
         self.docker_image_latest_tag = self.get_docker_image_latest_tag(self.docker_image_name, self.yml_docker_image)
+        if not self.yml_docker_image:
+            error_message, error_code = Errors.docker_image_does_not_exist_in_yml_file(self.file_path)
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                self.is_valid = False
         if not self.docker_image_latest_tag:
             error_message, error_code = Errors.non_existing_docker(self.yml_docker_image)
             if self.handle_error(error_message, error_code, file_path=self.file_path):